Move socket matching to util.net.
